+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## [0.1.7]
+* Move recentFoldersFile to User Data Dir.
+* Add Latest Project and All Project, sorted alphabetically.
+
 ## [0.1.6]
-   Add screenshots to README.md
-   Update README.md
+*  Add screenshots to README.md
+*  Update README.md
 
 ## [0.1.5]